Moved ashley based ai into own package
