+++ src/main/kotlin/eater/ai/ashley/DoIHaveThisComponentConsideration.kt
@@ -1,9 +1,8 @@
-package eater.ai
+package eater.ai.ashley
 
 import com.badlogic.ashley.core.Component
 import com.badlogic.ashley.core.ComponentMapper
 import com.badlogic.ashley.core.Entity
-import com.badlogic.gdx.math.MathUtils
 import kotlin.reflect.KClass
 
 class DoIHaveThisComponentConsideration<ToCheck: Component>(name:String, toCheck: KClass<ToCheck>): Consideration(name) {
