+++ src/main/kotlin/eater/ai/ashley/Extensions.kt
@@ -1,4 +1,4 @@
-package eater.ai
+package eater.ai.ashley
 
 import com.badlogic.gdx.math.MathUtils
 import com.badlogic.gdx.math.Vector2
