+++ src/main/kotlin/eater/ecs/systems/UpdateActionsSystem.kt
@@ -2,8 +2,7 @@ package eater.ecs.systems
 
 import com.badlogic.ashley.core.Entity
 import com.badlogic.ashley.systems.IntervalIteratingSystem
-import com.badlogic.ashley.systems.IteratingSystem
-import eater.ai.AiComponent
+import eater.ai.ashley.AiComponent
 import ktx.ashley.allOf
 
 class UpdateActionsSystem : IntervalIteratingSystem(allOf(AiComponent::class).get(), 0.03f) {
